changed the numPars int for a float, in order to use a ref to it with ofxUI

+++ src/ofxParticleEmitter.h
@@ -44,7 +44,7 @@ class ofxParticleEmitter
         float life;
         float lifeSpread;
         
-        int numPars;
+        float numPars;
         
         ofColor color;
         ofColor colorSpread;
